Separate test works more reliant
